feat(cli): urfave/cli skeleton and version subcommand
